A few changes
HTML:
Added races and titles to the HTML file.

CSS:
Added button style to the CSS style.

Javascript:
Added button states (mostly for users with slow devices/connection).
Added a race array along with a change to how the names are managed
(each race has an array, containing 2 arrays, one for each gender).
Added more properties to the character object.
Added a generator object, which contains the classes for the different
parts of the html page.
Added a function, that populates the race selection with options based
on the race array.
Added a new array containing the values from the form, using the new
generator object.